rename flatcoord -> ragged_array files
