+++ shapely/_ragged_array.py
